The proposed system can meet the objectives and it is obvious that the tourists and the travelers will be helped by this system.
The travel industry has experienced a very fast development of top-notch nature in the course of the most recent twenty years. This has basically been because of the coming of a fringeless' reality and expanded data spread about lofty situations all through the world. The expanding number of vacationers worldwide has fed the monetary and social effect extraordinarily. Be that as it may, the absence of data, correspondence, direction, arranging, and security issues has kept the travelers to skip visiting anyplace in Bangladesh and outside the nation. Along these lines, it has turned into an incredible part in Bangladesh to put with a business thought in a manner by helping voyagers by telling them about the spots, touring plans, costings, and, security. Android is the most used mobile operating system in the world right now, so design a system for android based to solve the problems faced by travelers is the best choice. Prohelika is an online travel management system application which is android based. This proposed system aimed to offer user so that they can have related information for a perfect trip, update about tour events from the group, also can have a view of cost estimation depending on the number of tourists, notification for upcoming trips. A user can have the direction to reach the destination like google play with the approximate budget to reach the place. Therefore this application is a trustworthy, secure and free of cost medium for an easy, comfortable and enjoyable travel experience. To solve the problems faced by travelers, inspire tourism by providing information, and enhancing people’s interest in traveling this android application Prohelika should create a positive outlook and gladly accepted by travelers. The project is divided into two module admin and user. We use Java, XML, Android Studio and Firebase database for the front end and back end design. After finishing the implementation, the proposed system can meet the objectives and it is obvious that the tourists and the travelers will be helped by this system.